Fix interactive calls, feedback messages
Attempt to call menu functions interactively and, if this fails due to
a wrong-type-error ("expected commandp"), then call the function
non-interactively. This now supports menu configurations that contain
both interactive or non-interactive menu item functions.

This patch also fixes the issue where "Menu aborted" would be printed in
the minibuffer message area when the menu function completed, regardless
of whether it was actually aborted or not.

Now the message "Menu ended." will be printed when the menu ends,
provided that another message is not already displayed (if, for example,
the selected menu function generated output).

When a menu is quit, a zero-length string is printed to effectively
erase the menu display.